refactor: extract AudioRecorderTypes.swift from AudioRecorder.swift
Extracts 4 supporting declarations (RecordedAudio struct, AVAudioRecorder
extension, AudioRecorderEngineFactory typealias, AudioRecorderCaptureFormat
enum) from AudioRecorder.swift into a new AudioRecorderTypes.swift.

Byte-identical move; SHA-256 043bc263e18e16b7efeb7eb300ddf83dc587953c900b50a6ee3e8a09aae0e55d
verified against origin/main:AudioRecorder.swift lines 5-58.

Imports: @preconcurrency AVFoundation (for AVAudioRecorder, AVFormatIDKey, etc.)
+ Foundation.

Reduces AudioRecorder.swift from 495 lines to 442 lines.

Tests: 667 (unchanged).
